requirements.txt
pyTelegramBotAPI
python-dotenv
requests
mysql-connector-python